Smith: rename the main executable project
