+++ src/Smith/Registry.cs
@@ -1,5 +1,4 @@
 using System.IO;
-using System.Reactive.Disposables;
 using System.Threading.Tasks;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Caching.Memory;
@@ -7,8 +6,6 @@ using Microsoft.Extensions.Options;
 using Splat;
 using TdLib;
 using Tel.Egram.Application;
-using Tel.Egram.Model.Messenger.Catalog;
-using Tel.Egram.Model.Messenger.Explorer;
 using Tel.Egram.Model.Messenger.Explorer.Factories;
 using Tel.Egram.Model.Notifications;
 using Tel.Egram.Model.Popups;
@@ -28,7 +25,7 @@ using Tel.Egram.Services.Utils.TdLib;
 using IBitmapLoader = Tel.Egram.Services.Graphics.IBitmapLoader;
 using BitmapLoader = Tel.Egram.Services.Graphics.BitmapLoader;
 
-namespace Tel.Egram
+namespace Smith
 {
     public static class Registry
     {
