+++ src/Smith/Smith.csproj
